apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lexai-api"
version = "0.1.0"
description = "LexAI — FastAPI chat API (Gemini, Supabase JWT, Redis rate limit, asyncpg, RAG)."
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.4.0",
  "pyjwt[crypto]>=2.9.0",
  "cryptography>=42.0.0",
  "asyncpg>=0.29.0",
  "redis>=5.0.0",
  "google-genai>=1.0.0",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["lexai*"]
